Handling better background image, adding return button, menu, handling better closing activity
