+++ app/src/main/AndroidManifest.xml
@@ -16,10 +16,21 @@
         <activity
             android:name=".MainActivity"
             android:exported="true">
+
             <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
+                <action android:name="android.intent.action.MAIN"/>
+                <action android:name="main.werewolfkotlin.MainActivity"/>
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+
+
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+        <activity
+            android:name=".SelectActivity"
+            android:exported="true">
 
             <meta-data
                 android:name="android.app.lib_name"
